Fixed hex result and added unit tests

+++ TransactionSearch.Core/Models/TransactionResponseDto.cs
@@ -10,11 +10,11 @@ namespace TransactionSearch.Core.Models
     public class TransactionData
     {
         public string BlockHash { get; set; }
-        public long BlockNumber { get; set; }
+        public double BlockNumber { get; set; }
         public string From { get; set; }
-        public long Gas { get; set; }
+        public decimal Gas { get; set; }
         public string Hash { get; set; }
         public string To { get; set; }
-        public long Value { get; set; }
+        public decimal Value { get; set; }
     }
 }
